# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ProfileDetail": {
        "properties": {
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/ProfileAnalysisRef"
            },
            "title": "Analyses",
            "type": "array"
          },
          "analyses_count": {
            "title": "Analyses Count",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/ProfileUserRef"
            },
            "title": "Users",
            "type": "array"
          },
          "users_count": {
            "title": "Users Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_active",
          "users_count",
          "analyses_count",
          "created_at",
          "updated_at",
          "users",
          "analyses"
        ],
        "title": "ProfileDetail",
        "type": "object"
      },
...
      "UserIdsBody": {
        "example": {
          "user_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ]
        },
        "properties": {
          "user_ids": {
            "description": "Conjunto completo de usuários do perfil; substitui o atual.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "User Ids",
            "type": "array"
          }
        },
        "required": [
          "user_ids"
        ],
        "title": "UserIdsBody",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "description": "Token opaco emitido por `POST /auth/token`. Cole apenas o token (o Swagger acrescenta o prefixo `Bearer`).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API HTTP do servidor de análise de dados via MCP: emissão de tokens (`/auth`), dados do próprio\nusuário (`/me`), administração (`/admin/*`) e `/health`.\n\n**Autenticação.** Emita um token em `POST /auth/token` (e-mail + senha), clique em **Authorize**\ne cole só o token. `/me` exige um token válido; `/admin/*` exige também administrador.\n\n**Erros.** As rotas de `/auth`, `/me` e `/admin` respondem erros de domínio no formato\n`{\"error\": \"<slug>\", \"message\": \"...\"}`. Erros de validação do corpo ou dos parâmetros vêm no\nformato padrão do FastAPI (`422`, `{\"detail\": [...]}`).\n\n**O `/mcp` não está neste documento.** O endpoint MCP (Streamable HTTP/JSON-RPC) é descrito em\n`docs/MCP.md`.\n",
    "title": "Análise de Dados Genérica com MCP — API HTTP",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/profiles/{profile_id}/users": {
      "put": {
        "description": "Substitui o conjunto de usuários do perfil (lista vazia remove todos).",
        "operationId": "set_profile_users_admin_profiles__profile_id__users_put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `profile_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Set Profile Users Admin Profiles  Profile Id  Users Put"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Definir usuários do perfil",
        "tags": [
          "admin-profiles"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "description": "Lista paginada, com busca textual (`q`, nome ou e-mail) e filtros por bloqueio e perfil.",
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "is_blocked",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Blocked"
            }
          },
          {
            "in": "query",
            "name": "profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Listar usuários",
        "tags": [
          "admin-users"
        ]
      },
      "post": {
        "description": "Cria o usuário com senha inicial (política: mínimo 6 caracteres, maiúscula, minúscula, número e símbolo) e vincula os perfis informados, numa transação.",
        "operationId": "create_user_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Requisição inválida. Slugs: `invalid_password`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `email_already_exists`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Create User Admin Users Post"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Criar usuário",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "description": "Exclusão física, só para usuário sem histórico de execuções; com histórico responde 409 e o caminho é bloquear (`POST /admin/users/{user_id}/block`). Não é possível excluir a si mesmo nem o último administrador ativo.",
        "operationId": "delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `self_protected`, `last_admin_protected`, `user_has_history`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Excluir usuário",
        "tags": [
          "admin-users"
        ]
      },
      "get": {
        "description": "Devolve o usuário com perfis e quantidade de tokens ativos. Nunca devolve hash de senha.",
        "operationId": "get_user_admin_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Detalhar usuário",
        "tags": [
          "admin-users"
        ]
      },
      "patch": {
        "description": "Altera apenas os campos enviados (ao menos um). O administrador não pode remover o próprio papel, e o último administrador ativo não pode deixar de sê-lo.",
        "operationId": "update_user_admin_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `email_already_exists`, `self_protected`, `last_admin_protected`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Alterar usuário",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/{user_id}/block": {
      "post": {
        "description": "Bloqueia o usuário: ele perde o acesso na chamada seguinte. Não é possível bloquear a si mesmo nem o último administrador ativo.",
        "operationId": "block_user_admin_users__user_id__block_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `self_protected`, `last_admin_protected`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Bloquear usuário",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/{user_id}/password": {
      "put": {
        "description": "Define uma nova senha para o usuário (política da criação). Revoga os tokens ativos dele.",
        "operationId": "reset_password_admin_users__user_id__password_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Requisição inválida. Slugs: `invalid_password`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Redefinir senha",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/{user_id}/profiles": {
      "put": {
        "description": "Substitui o conjunto de perfis do usuário (lista vazia remove todos).",
        "operationId": "set_user_profiles_admin_users__user_id__profiles_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIdsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Set User Profiles Admin Users  User Id  Profiles Put"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Definir perfis do usuário",
        "tags": [
          "admin-users"
        ]
      }
    },
...
  },
...
}